+++ forui/example/lib/sandbox.dart
@@ -1,29 +1,7 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
 import 'package:forui/forui.dart';
 
-const features = ['Keyboard navigation', 'Typeahead suggestions', 'Tab to complete'];
-
 const fruits = ['Apple', 'Banana', 'Orange', 'Grape', 'Strawberry', 'Pineapple'];
-
-const letters = {
-  'A': 'A',
-  'B': 'B',
-  'C': 'C',
-  'D': 'D',
-  'E': 'E',
-  'F': 'F',
-  'G': 'G',
-  'H': 'H',
-  'I': 'I',
-  'J': 'J',
-  'K': 'K',
-  'L': 'L',
-  'M': 'M',
-  'N': 'N',
-  'O': 'O',
-};
 
 class Sandbox extends StatefulWidget {
   const Sandbox({super.key});
@@ -32,27 +10,8 @@ class Sandbox extends StatefulWidget {
   State<Sandbox> createState() => _SandboxState();
 }
 
-enum Notification { all, direct, nothing, limitedTime, timeSensitive, selectedApps }
-
-class _SandboxState extends State<Sandbox> with SingleTickerProviderStateMixin {
+class _SandboxState extends State<Sandbox> {
   @override
-  Widget build(BuildContext context) => Center(
-    child: Column(
-      mainAxisSize: .min,
-      children: [
-        FTimeField(),
-        FDateField(),
-        FAutocomplete(items: fruits),
-        FSelect<String>(items: {for (final fruit in fruits) fruit: fruit}),
-        FMultiSelect<String>(items: {for (final fruit in fruits) fruit: fruit}),
-        FTextField.password(),
-        FTextField(
-          suffixBuilder: (_, style, states) => IconTheme(
-            data: style.clearButtonStyle.iconContentStyle.iconStyle.resolve(states),
-            child: Icon(Icons.search),
-          ),
-        ),
-      ],
-    ),
-  );
+  Widget build(BuildContext context) =>
+      Center(child: FSelect<String>.search(items: {for (final fruit in fruits) fruit: fruit}));
 }