+++ forui/test/src/widgets/select/multi/multi_select_golden_test.dart
@@ -117,7 +117,9 @@ void main() {
           child: FMultiSelect<String>.searchBuilder(
             key: key,
             format: Text.new,
-            style: TestScaffold.blueScreen.multiSelectStyle,
+            style: TestScaffold.blueScreen.multiSelectStyle.copyWith(
+              searchStyle: const .delta(fieldStyle: .delta(cursorColor: Color(0xFF03A9F4))),
+            ),
             contentScrollHandles: true,
             filter: (_) => [],
             contentBuilder: (_, _, _) => [for (int i = 0; i < 10; i++) .item(title: Text('$i'), value: '$i')],
@@ -137,7 +139,9 @@ void main() {
           child: FMultiSelect<String>.searchBuilder(
             key: key,
             format: Text.new,
-            style: TestScaffold.blueScreen.multiSelectStyle,
+            style: TestScaffold.blueScreen.multiSelectStyle.copyWith(
+              searchStyle: const .delta(fieldStyle: .delta(cursorColor: Color(0xFF03A9F4))),
+            ),
             contentScrollHandles: true,
             filter: (_) async {
               await Future.delayed(const Duration(seconds: 1));
@@ -160,7 +164,9 @@ void main() {
           child: FMultiSelect<int>.search(
             {for (int i = 0; i < 10; i++) '$i': i},
             key: key,
-            style: TestScaffold.blueScreen.multiSelectStyle,
+            style: TestScaffold.blueScreen.multiSelectStyle.copyWith(
+              searchStyle: const .delta(fieldStyle: .delta(cursorColor: Color(0xFF03A9F4))),
+            ),
             contentScrollHandles: true,
             filter: (_) => [],
           ),